+++ test/core-options-verify/verifySkipIfNewer.test.js
@@ -52,7 +52,7 @@ test('verifySkipIfNewer/undefined', () => {
   expect(w.verify(expectedWarn)).resolves.toBe(true)
 })
 
-test.each([null, ''])('verifySkipIfNewer/blank', (value) => {
+test.each(dt.BlankValueDataSet)('verifySkipIfNewer/blank', (value) => {
   const workingObject = {
     userOption: { skipIfNewer: value },
     validatedOption: {}
@@ -98,7 +98,7 @@ test('verifySkipIfNewer/undefined/quiet', () => {
     })
   expect(co0.warn).not.toBeCalled()
 })
-test.each([null, ''])('verifySkipIfNewer/blank/quiet', (value) => {
+test.each(dt.BlankValueDataSet)('verifySkipIfNewer/blank/quiet', (value) => {
   const workingObject = {
     userOption: { skipIfNewer: value },
     validatedOption: { quiet: true }